extension/backend/requirements.txt
# UnReal ML Backend Requirements
# Install with: pip install -r requirements.txt

# Image Detection Backend (FastAPI)
fastapi>=0.109.0
uvicorn>=0.27.0

# Text Detection Backend (Flask)
flask>=3.0.0
flask-cors>=4.0.0

# Shared ML Dependencies
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
pydantic>=2.0.0

# Video Analysis Dependencies
opencv-python>=4.8.0
yt-dlp>=2024.1.0
numpy>=1.24.0

# Fine-tuning Dependencies
datasets>=2.16.0
scikit-learn>=1.3.0
accelerate>=0.25.0
tqdm>=4.66.0

# Audio Detection Dependencies
librosa>=0.10.0
soundfile>=0.12.0

# Async HTTP client (for inter-service communication)
httpx>=0.26.0
